# bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "ISI-WebAPI, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/Funcionario": {
      "get": {
        "tags": [
          "Funcionario"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/FuncionarioModelListServiceResponse"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Funcionario"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/FuncionarioModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/FuncionarioModelListServiceResponse"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Funcionario"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/FuncionarioModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/FuncionarioModelListServiceResponse"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Funcionario"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/FuncionarioModelListServiceResponse"
            }
          }
        }
      }
    },
    "/api/Funcionario/{id}": {
      "get": {
        "tags": [
          "Funcionario"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/FuncionarioModelServiceResponse"
            }
          }
        }
      }
    },
    "/api/Funcionario/inativaFuncionario": {
      "put": {
        "tags": [
          "Funcionario"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/FuncionarioModelListServiceResponse"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "DepartamentoEnum": {
      "enum": [
        "RH",
        "IT",
        "Comercial",
        "Design",
        "Acessórios"
      ],
      "type": "string"
    },
    "FuncionarioModel": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "nome": {
          "type": "string"
        },
        "apelido": {
          "type": "string"
        },
        "departamento": {
          "$ref": "#/definitions/DepartamentoEnum"
        },
        "ativo": {
          "type": "boolean"
        },
        "turno": {
          "$ref": "#/definitions/TurnoEnum"
        },
        "dataCriacao": {
          "format": "date-time",
          "type": "string"
        },
        "dataAlteracao": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
    "FuncionarioModelListServiceResponse": {
      "type": "object",
      "properties": {
        "dados": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/FuncionarioModel"
          }
        },
        "mensagem": {
          "type": "string"
        },
        "sucesso": {
          "type": "boolean"
        }
      }
    },
    "FuncionarioModelServiceResponse": {
      "type": "object",
      "properties": {
        "dados": {
          "$ref": "#/definitions/FuncionarioModel"
        },
        "mensagem": {
          "type": "string"
        },
        "sucesso": {
          "type": "boolean"
        }
      }
    },
    "TurnoEnum": {
      "enum": [
        "Manhã",
        "Normal",
        "Noite"
      ],
      "type": "string"
    }
  }
}